Move tls code to its own file
